backend/requirements.txt
Django==3.2
django-colorfield==0.8.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.8.6
python-dotenv==1.0.0
requests==2.28.2
sorl-thumbnail==12.9.0
